requirements.txt
web3==5.2.2
Django==2.1.7
django-pyodbc-azure==2.1.0.0
pyodbc==4.0.24
pytz==2018.7
djangorestframework==3.10.3
django-filter==2.0.0
django-configurations==2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
solc==3.2.0